Move shared components and fix small errors
