+++ content/resources/dev-tools/components/LogEntry.tsx
